Separated convert clock from timer file
